refactor: :recycle: refactor rounding function
Moving the rounding algorithm function to separate file.
